--- src/components/Page/CardAddPage.js
@@ -1,258 +1,0 @@
-import React, { useEffect, useRef, useState } from 'react';
-import BackButton from '../BackButton/BackButton';
-import Card from '../Card/Card';
-import Input from '../Input/Input';
-import InputContainer from '../InputContainer/InputContainer';
-import ModalPage from './ModalPage';
-import TextButton from '../TextButton/TextButton';
-import Tooltip from '../Tooltip/Tooltip';
-import CardRegisterPage from './CardRegisterPage';
-
-const CardAddPage = (props) => {
-  const [cardCompany, setCardCompany] = useState({ name: '', color: '' });
-  const [expiration, setExpiration] = useState({ month: '', year: '' });
-  const [ownerName, setOwnerName] = useState('');
-  const [securityCode, setSecurityCode] = useState('');
-  const [password, setPassword] = useState({
-    first: '',
-    second: '',
-  });
-  const [cardNumbers, setCardNumbers] = useState({
-    first: '',
-    second: '',
-    third: '',
-    fourth: '',
-  });
-  const [isModalOpened, setIsModalOpened] = useState(false);
-  const [isAllValid, setIsAllValid] = useState(false);
-
-  const resetState = () => {
-    setCardCompany({ name: '', color: '' });
-    setExpiration({ month: '', year: '' });
-    setOwnerName('');
-    setSecurityCode('');
-    setPassword({
-      first: '',
-      second: '',
-    });
-    setCardNumbers({
-      first: '',
-      second: '',
-      third: '',
-      fourth: '',
-    });
-    setIsModalOpened(false);
-    setIsAllValid(false);
-  };
-
-  const handleCardCompany = ({ target }) => {
-    const company = target.closest('li').dataset.company;
-    setCardCompany({
-      name: `${company} 카드`,
-      color: `bg-${company}`,
-    });
-
-    setIsModalOpened(false);
-  };
-
-  const handleExpirationInput = ({ target: { value } }, category) => {
-    const valueAsString = String(value);
-
-    if (valueAsString.length > 2) {
-      return;
-    }
-
-    setExpiration({
-      ...expiration,
-      [category]: valueAsString.replace(/[^0-9]s/g, ''),
-    });
-  };
-
-  const handleCardNumbers = ({ target: { value } }, key) => {
-    const numberReg = /^[0-9]{1,4}$/gi;
-    if (!numberReg.test(Number(value))) {
-      return;
-    }
-
-    setCardNumbers({ ...cardNumbers, [key]: value });
-  };
-
-  useEffect(() => {
-    if (cardNumbers.first.length + cardNumbers.second.length === 8 && !cardCompany.name) {
-      setIsModalOpened(true);
-    }
-  }, [cardNumbers, cardCompany]);
-
-  const handleOwnerName = ({ target: { value } }) => {
-    setOwnerName(value.trimStart());
-  };
-
-  const handleSecurityCodeInput = ({ target: { value } }) => {
-    const numberReg = /^[0-9]{1,4}$/gi;
-
-    if (!numberReg.test(Number(value))) {
-      return;
-    }
-
-    setSecurityCode(value);
-  };
-
-  const handlePasswordInput = ({ target: { value, name } }) => {
-    const numberReg = /^[0-9]{1,4}$/gi;
-    if (!numberReg.test(Number(value))) {
-      return;
-    }
-
-    setPassword({ ...password, [name]: value });
-  };
-
-  const handleCardInfoSubmit = (e) => {
-    e.preventDefault();
-
-    setIsAllValid(true);
-  };
-
-  return !isAllValid ? (
-    <div className="p-5">
-      <div className="flex items-center">
-        <BackButton />
-        <h1 className="text-xl ml-4">카드 추가</h1>
-      </div>
-      <div className="flex  justify-center my-7">
-        <Card
-          name={ownerName || 'NAME'}
-          expiration={`${expiration.month || 'MM'}/${expiration.year || 'YY'}`}
-          cardCompany={cardCompany}
-          cardNumbers={cardNumbers}
-        />
-      </div>
-
-      <form onSubmit={handleCardInfoSubmit}>
-        <InputContainer title={'카드 번호'} bgColor={'bg-gray-250'} width={'w-full'}>
-          <>
-            {Array.from({ length: 4 }).map((_, index) => {
-              const currentKey = Object.keys(cardNumbers)[index];
-              return (
-                <React.Fragment key={`cardNumber-${currentKey}`}>
-                  <Input
-                    type={index > 1 ? 'password' : 'text'}
-                    length={4}
-                    placeholder={'0000'}
-                    value={cardNumbers[currentKey]}
-                    onChange={(e) => handleCardNumbers(e, currentKey)}
-                    className={'px-4'}
-                  />
-                  {index === 3 ? '' : <span>-</span>}
-                </React.Fragment>
-              );
-            })}
-          </>
-        </InputContainer>
-
-        <InputContainer title={'만료일'} width={'w-5/12'} bgColor={'bg-gray-250'}>
-          <>
-            <Input
-              placeholder={'MM'}
-              type={'number'}
-              min={1}
-              max={12}
-              value={expiration.month}
-              onChange={(e) => handleExpirationInput(e, 'month')}
-              className={'text-center pl-4'}
-            />
-            <span className="text-gray-400 ">/</span>
-            <Input
-              placeholder={'YY'}
-              type={'number'}
-              min={21}
-              value={expiration.year}
-              onChange={(e) => handleExpirationInput(e, 'year')}
-              className={'text-center pr-4'}
-            />
-          </>
-        </InputContainer>
-
-        <InputContainer
-          title={'카드 소유자 이름(선택)'}
-          bgColor={'bg-gray-250'}
-          count={ownerName.length}
-          width={'w-full'}
-        >
-          <Input
-            minLength={0}
-            length={30}
-            placeholder={'카드에 표시된 이름과 동일하게 입력하세요.'}
-            value={ownerName}
-            onChange={handleOwnerName}
-            required={false}
-            className={'px-4'}
-          />
-        </InputContainer>
-
-        <InputContainer title={'보안코드(CVC/CVV)'} width={'w-1/3'}>
-          <>
-            <Input
-              type={'password'}
-              length={3}
-              value={securityCode}
-              className={'px-4 text-center'}
-              onChange={handleSecurityCodeInput}
-            />
-            <Tooltip content={<img src="images/CVC.png" alt="카드 뒷면 서명란 끝의 세자리 숫자입니다." />} />
-          </>
-        </InputContainer>
-
-        <InputContainer title={'카드 비밀번호'} width={'w-7/12'}>
-          <>
-            <Input
-              className={'mr-1.5 text-center'}
-              width={'small'}
-              type={'password'}
-              length={1}
-              value={password.first}
-              name={'first'}
-              onChange={handlePasswordInput}
-            />
-            <Input
-              className={'mr-1.5 text-center'}
-              width={'small'}
-              type={'password'}
-              length={1}
-              value={password.second}
-              name={'second'}
-              onChange={handlePasswordInput}
-            />
-            <Input
-              className={'mr-1.5 bg-opacity-0 text-center'}
-              width={'small'}
-              type={'password'}
-              length={1}
-              value={'*'}
-              disabled={true}
-            />
-            <Input
-              className={'mr-1.5 bg-opacity-0 text-center'}
-              width={'small'}
-              type={'password'}
-              length={1}
-              value={'*'}
-              disabled={true}
-            />
-          </>
-        </InputContainer>
-        <TextButton text={'다음'} />
-      </form>
-
-      {isModalOpened && <ModalPage onClick={handleCardCompany} />}
-    </div>
-  ) : (
-    <CardRegisterPage
-      cardCompany={cardCompany}
-      cardNumbers={cardNumbers}
-      expiration={expiration}
-      ownerName={ownerName}
-      resetState={resetState}
-    />
-  );
-};
-export default CardAddPage;
